chore(cdr): Move httpapi to /coderd directory
httpapi is specific to coderd and should be scoped under coderd
