+++ coderd/coderd.go
@@ -11,7 +11,7 @@ import (
 
 	"cdr.dev/slog"
 	"github.com/coder/coder/database"
-	"github.com/coder/coder/httpapi"
+	"github.com/coder/coder/coderd/httpapi"
 	"github.com/coder/coder/coderd/httpmw"
 	"github.com/coder/coder/site"
 )
