+++ coderd/parameters.go
@@ -12,7 +12,7 @@ import (
 
 	"github.com/coder/coder/codersdk"
 	"github.com/coder/coder/database"
-	"github.com/coder/coder/httpapi"
+	"github.com/coder/coder/coderd/httpapi"
 )
 
 func (api *api) postParameter(rw http.ResponseWriter, r *http.Request) {
