+++ coderd/provisionerjobs.go
@@ -17,7 +17,7 @@ import (
 
 	"github.com/coder/coder/codersdk"
 	"github.com/coder/coder/database"
-	"github.com/coder/coder/httpapi"
+	"github.com/coder/coder/coderd/httpapi"
 )
 
 // Returns provisioner logs based on query parameters.
