+++ codersdk/client.go
@@ -14,7 +14,7 @@ import (
 	"golang.org/x/xerrors"
 
 	"github.com/coder/coder/coderd/httpmw"
-	"github.com/coder/coder/httpapi"
+	"github.com/coder/coder/coderd/httpapi"
 )
 
 // New creates a Coder client for the provided URL.
